requirements.txt
apache-airflow[amazon]==2.9.1
boto3==1.34.84
botocore==1.34.84
pyspark==3.5.1
